sound is still having some issues, i think there may be bugs in aframe
a-frame triangle is being called to a specific class in the javascript, however a-frame circle is playing instead. i have troubleshooted on the console and it says circle is playing, but there is no link to circle.  i thinkn maybe the shapes in aframe are all coming under the same parent so when they are called they're all reacting.